Extract exec functions to a module
